refactor: ♻️ reestruturação da tela Home na aplicação
